session_supervisor: carve launch.rs (pure move)
